+++ src/mapping.js
@@ -4,4 +4,7 @@ export const mapping = {
   quotation: process.env.NODE_ENV === 'development'
     ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/quotation'
     : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/quotation',
+  trackVisit: process.env.NODE_ENV === 'development'
+    ? 'http://127.0.0.1:5001/carllouismanuel-1e3a9/us-central1/trackVisit'
+    : 'https://us-central1-carllouismanuel-1e3a9.cloudfunctions.net/trackVisit',
 }